% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abntools.R
\docType{package}
\name{abntools}
\alias{abntools}
\title{abntools: A package for working with ABNs and ACNs}
\description{
The abntools package provides tools to work with Australian Business Numbers
(ABNs) and Australian Company Numbers (ACNs). The most important of these
tools is an interface with the ABN lookup API.
}
\section{Validation functions}{

The functions \code{is_abn} and \code{is_acn} are provided to determine if an
input is in the format of a valid ABN or ACN respectively.
}

