% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_abn.R
\name{is_abn}
\alias{is_abn}
\title{Use a checksum to identify if a number is in the format of an ABN.}
\usage{
is_abn(x)
}
\arguments{
\item{x}{A numeric or character vector of numbers to check}
}
\value{
A logical vector of the same length as \code{x} identifying if
each element is in the format of a valid ABN
}
\description{
This function identifies if a number is in the format of a valid ABN, using
the rules oultined \href{https://abr.business.gov.au/Help/AbnFormat}{here}. Note
that this function does not confirm that the ABN is valid; it simply checks
if it \emph{could} be valid.
}
\examples{
# Single ABN
is_abn(19621994018)

# Numeric vector
is_abn(c(19621994018, 81601798609))

# Character vector
is_abn(c("50 110 219 460", "79 769 424 861"))
}
