% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_acn.R
\name{is_acn}
\alias{is_acn}
\title{Use a checksum to identify if a given input is in the format of an ACN}
\usage{
is_acn(x)
}
\arguments{
\item{x}{A character vector of ACNs to check}
}
\value{
A logical vector of the same length as \code{x} identifying if each
element is in the format of a valid ACN
}
\description{
This function identifies if an input is in the format of a valid ACN using
the rules oultined \href{https://asic.gov.au/for-business/registering-a-company/steps-to-register-a-company/australian-company-numbers/australian-company-number-digit-check/}{here}. Note
that this function does not confirm that the ACN is valid or active; it
simply checks if it \emph{could} be valid.
}
\examples{
# Single ACN
is_acn("110219460")

# Character vector which may include arbitrary white space
is_acn(c("001 250 004", "001 999 999"))
}
