% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_abn.R
\name{is_abn}
\alias{is_abn}
\title{Use a checksum to identify if an ABN is valid .}
\usage{
is_abn(abns)
}
\arguments{
\item{abns}{A numeric or character vector of ABNs}
}
\value{
A logical vector of the same length as abns identifying if each
element is a valid ABN
}
\description{
Use a checksum to identify if an ABN is valid .
}
\examples{
    is_abn(19621994018)
}
